pyproject.toml
[tool.poetry]
name = "smikic-dwh"
version = "0.1.0"
description = "Personal DWH"
authors = ["stejul <stejul@rocketmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7.1"
SQLAlchemy = "^1.4.15"
luigi = "^3.0.3"
python-dotenv = "^0.17.1"
kafka-python = "^2.0.2"
requests = "^2.25.1"
tweepy = "^3.10.0"
Jinja2 = "^3.0.1"
pymongo = "^3.11.4"
pandas = "^1.2.4"
psycopg2-binary = "^2.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
